# backend/openapi/Castor.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Castor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/grant-calls": {
      "get": {
        "tags": [
          "AdminGrantCalls"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrantCallResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrantCallResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrantCallResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminGrantCalls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantCallRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantCallRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grant-calls/draft": {
      "post": {
        "tags": [
          "AdminGrantCalls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftGrantCallRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftGrantCallRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DraftGrantCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grant-calls/{grantCallId}": {
      "put": {
        "tags": [
          "AdminGrantCalls"
        ],
        "parameters": [
          {
            "name": "grantCallId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseGrantCallRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseGrantCallRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseGrantCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grant-calls/{grantCallId}/open": {
      "post": {
        "tags": [
          "AdminGrantCalls"
        ],
        "parameters": [
          {
            "name": "grantCallId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grant-calls/{grantCallId}/close": {
      "post": {
        "tags": [
          "AdminGrantCalls"
        ],
        "parameters": [
          {
            "name": "grantCallId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCallResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/innovations/{innovationId}/fit": {
      "post": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFitAssessmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFitAssessmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFitAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "teryt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/innovations/{innovationId}/fit/{fitAssessmentId}": {
      "get": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fitAssessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/innovations/{innovationId}/fit/{fitAssessmentId}/recalculate": {
      "post": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fitAssessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitAssessmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/innovations/{innovationId}/fit/{fitAssessmentId}/assistant": {
      "get": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fitAssessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FitAssessments"
        ],
        "parameters": [
          {
            "name": "innovationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fitAssessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskFitAssistantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AskFitAssistantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AskFitAssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FitAssistantMessageResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AskFitAssistantRequest": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateFitAssessmentRequest": {
        "required": [
          "teryt"
        ],
        "type": "object",
        "properties": {
          "teryt": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateGrantCallRequest": {
        "required": [
          "title",
          "description",
          "criteria",
          "challengeAreaCodes",
          "opensOn",
          "closesOn"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "criteria": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "challengeAreaCodes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "opensOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "closesOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
...
      "DraftGrantCallRequest": {
        "required": [
          "challengeAreaCode",
          "teryt"
        ],
        "type": "object",
        "properties": {
          "challengeAreaCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "teryt": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FitAssessmentMunicipalityResponse": {
        "required": [
          "teryt",
          "name",
          "powiat"
        ],
        "type": "object",
        "properties": {
          "teryt": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "powiat": {
            "type": "string"
          }
        }
      },
      "FitAssessmentResponse": {
        "required": [
          "id",
          "innovationId",
          "innovationTitle",
          "municipality",
          "dataYear",
          "fit",
          "summary",
          "unchanged",
          "toAdapt",
          "missing",
          "serviceProvider",
          "serviceForm",
          "scaleEstimate",
          "comparison",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "innovationId": {
            "type": "string",
            "format": "uuid"
          },
          "innovationTitle": {
            "type": "string"
          },
          "municipality": {
            "$ref": "#/components/schemas/FitAssessmentMunicipalityResponse"
          },
          "dataYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fit": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "unchanged": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "toAdapt": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "serviceProvider": {
            "type": [
              "null",
              "string"
            ]
          },
          "serviceForm": {
            "type": [
              "null",
              "string"
            ]
          },
          "scaleEstimate": {
            "type": [
              "null",
              "string"
            ]
          },
          "comparison": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FitComparisonRowResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FitAssistantMessageResponse": {
        "required": [
          "id",
          "role",
          "text",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FitComparisonRowResponse": {
        "required": [
          "requirement",
          "indicatorName",
          "unit",
          "level",
          "value",
          "regionAverage",
          "year"
        ],
        "type": "object",
        "properties": {
          "requirement": {
            "type": "string"
          },
          "indicatorName": {
            "type": "string"
          },
          "unit": {
            "type": [
              "null",
              "string"
            ]
          },
          "level": {
            "type": "string"
          },
          "value": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "regionAverage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "GrantCallResponse": {
        "required": [
          "id",
          "title",
          "description",
          "criteria",
          "challengeAreaCodes",
          "opensOn",
          "closesOn",
          "status",
          "isOpen",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "criteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "challengeAreaCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "opensOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "closesOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "isOpen": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReviseGrantCallRequest": {
        "required": [
          "title",
          "description",
          "criteria",
          "challengeAreaCodes",
          "opensOn",
          "closesOn"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "criteria": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "challengeAreaCodes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "opensOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "closesOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
...
    }
  },
...
}